short int

+++ src/SolarSystem.cpp
@@ -36,7 +36,7 @@ void SolarSystem::orbit(size_t days) {
     
     for (size_t i = 0; i < days; i++) {
         for (auto& planet : map_) {
-            int slower = 1;
+            short int slower = 1;
             auto newX = static_cast<double>(planet.getDistance() * cos(angle/slower));
             auto newY = static_cast<double>(planet.getDistance() * sin(angle/slower));
             planet.setPos(newX, newY);
